Fixed alsa_default.cpp calls to streamName() that parameter.
The calls in the LOGV(...) calls using streamName() needed to have
a parameter for an alsa_handle_t "handle" passed in. This is in the
upstream AOSP Google code so this will need to be pushed upstream if
we want to fix the use of LOGV messages and have this compile properly.

Also fixed miss used "stream" pointer in location that should have used
streamName() function call.

Change-Id: I8cfa51a6a9d9f44d6b6bfe00978a2d7e26296a5f