[PORT FROM R2] [AUDIO] Randomly tone not heard on HDMI
BZ: 17287

While HDMI connected, and dialing, some tone are not heard.
For each key press, the setParameter fonction will be called on the output
stream. It will lead to route it (unroute from previous route = closing alsa device
then routing the new route = opening alsa device).
If the route is the same, and the conditions are identical (ie same mode, same device,
stream already attached to the same route), this patch will bypass the unroute/route.

Change-Id: Iea708ec39452eb969b51653ca4bcc7dcabb78e83
Old-Change-Id: I470f8285e4d894a6a6581249831cbb4749a93122
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/27789
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>