[PORT FROM R2][AUDIO] Audio Glitches at the establishment/release of voice call
BZ: 17698

Glitch are listened at the establishement/release of call.
It was rootcaused that at initialisation of a call, the modem configures
its port by himself (AMC block) concurrently to PW processor.

It was etablished that we have to open the audio route only after the
modem has sent the xcallstat = Active (MT) or alerting (MO).

As a result, HAL had to listen to this xcallstat unsollicited AT command
to enable/disable audio routing for a call

The purpose of this patch is to:

-register to ATManager as a listener of modem state.
-register to listen XCALLSTAT
-implement the observer of XCALLSTAT
-Implement a new state machine for route accessibility

Change-Id: I825d474539e5e0b3bc9f4bff03d1e4fc83f44b89
Orig-Change-Id: Iad37ee390cd1638f9dd3f825a7303190a9aefc38
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/30890
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>