Audio: Widi audio middleware support.
BZ: 19806

The Widi audio virtual device require additional support from the audio system.

Changes to support audio routing to the Widi virtual device when the Widi is on.
The audio routing will be enabled only when an intent, widi on/off, is received.
This code is tested by broadcasting an intent from command line.
This Patch depends on change #31592, 31594 and 31595.

Change-Id: I422ca1c1fd94dcf547b038a1b0cfd240cd38d7d6
Signed-off-by: Rajneesh Chowdhury <rajneeshx.chowdhury@intel.com>
Reviewed-on: http://android.intel.com:8080/31593
Reviewed-by: Marotte, Fabien <fabien.marotte@intel.com>
Reviewed-by: Gaffie, FrancoisX <francoisx.gaffie@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>