[PORT FROM R2][AUDIO] Over consumption in CSV call
BZ: 19461

An overconsumption has been observed, about 30 mW, which is exactly
the impact of enabling I2S2 from modem side.

The purpose of this patch is to correctly managed the enable/disable
of the I2S2 port in case of mixing AND record (same issue)

This patch deals with audio HAL. A playback stream is always opened
and put in standby to trig the establishment of a CSV call.
The standby function has to disable the I2S2 of the modem immediately.
The same has been added also for the record.

Change-Id: I2cf0444709df8fe163d7fdce0a015745c888040d
Old-Change-Id: I49f6ca64e192f54210f5b1fcc6ad3412bdd224cb
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/31892
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>