[PORT FROM R2][AUDIO] Noticeable gitches heard on the distant when switching from CSV to/from VoIP
BZ: 18343

When switching from/to CSV to/from VoIP, distant can hear noticeable salve of glitches.
While the HW is still not configured, a NULL source is used during transition states.
The NULL alsa plugin only supports MMAP access mode (even if not complaining when opened
in RW_INTERLEAVED mode). Standalone test on null devices in RW capture showed:
-duration is wrong
-periodic glitch heard

The same capture test in MMAP mode were reporting an exact duration and a perfect silence.

As HAL is designed to open in RW mode, and migrating to MMAP is too risky, this patch just
detects the type of plugin used. If NULL, the input just returns silence to the client.
(The output stream was changed as well for consistency)

Change-Id: Ic011a327bacdaeb0a7ca042c02ed909a23ed9357
Old-Change-Id: Idd9c37e7eaec409f9946d3955274f339261c6d05
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/33345
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>