[PORT FROM R2][AUDIO] Media server crashes while using record app from Android store
BZ: 21358

While using record apps from android store, media server crash has been
observed.

It is due to the fact that unrouting request (ie route to device null) was
not taken into account, since it was followed by a delete of the input stream.

It is now established that somme app might delay the delete of the input
stream and we have to be robust to this.

This patch implements the unroute while requesting a null device.

Change-Id: Ic2a9227d2b1196e3f2b7f38383bc3d514054164f
Old-Change-Id: I14dd24a3c575afad2869a6acf399476a3579c51b
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/33351
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>