[PORT FROM R2][AUDIO] No audio when calling voice mail on free network
BZ: 21343

No Audio when calling voice mail on free network since XPROGRESS status
is never received.
In order not to break what has been fixed before in chinese network and
to avoid glitches from the modem, the best compromise is to listen
both XPROGRESS and XCALLSTAT and to enable audio on the appropriate state
of XPROGRESS OR the appropriate state of XCALLSTAT

This patch implements this dual mechanism on HAL part.

Change-Id: I8dcbc167af93c09606380f85ced609eb79f6a783
Old-Change-Id: Ie2f722139624f078243fc978ebb224ef031925a0
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/33352
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>