[PORT FROM R2][AUDIO] Orange Visual Voice Mail application does not work
BZ: 22896

Orange Visual Voice Mail application does not render audio.
Checking at the logs, it appears that Audio HAL was generating silence
and the mode was "forced" to in call.

In call mode needs callstatus to enable the voice route.
In order to have audio in this case, the voice route is "derouted" to media
route as previously done for Modem Reset during call or at the end of a call.
The route deviation process is now extended.

Note that it will not work with BT as Audio HAL cannot use BT SCO in normal mode
and as it also cannot take decision to play on a different device.

Change-Id: I0abd6566efe17f7d19a5216523c03a9a8537d2fe
Orig-Change-Id: I936a6d94a00a40f367b35ac80668cb9574783469
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/35604
Reviewed-by: De Chivre, RenaudX <renaudx.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>