[Audio] Change return value of hasBackMicrophone
BZ: 23280

When Camcorder is launched, it should use back mic rather than front mic.
A test is done with the function hasBackMicrophone which returns always
the same value - true or false - depending on the HW.
It returned false, but on our HW there is a back mic, so we change the
value it returns to true.
Modification done in daughter class.

Change-Id: Icdcd7afe95c6519b9ac75e337357b24fa9646bae
Signed-off-by: Pierre-LouisX Tonnet <pierre-louisx.tonnet@intel.com>
Reviewed-on: http://android.intel.com:8080/35460
Reviewed-by: Afonso, PhilippeX <philippex.afonso@intel.com>
Reviewed-by: De Chivre, RenaudX <renaudx.de.chivre@intel.com>
Reviewed-by: Bellouard, Jean-marieX <jean-mariex.bellouard@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>