[AUDIO] Record in mode normal does not work when modem is not available
BZ: 25820

On some CTP board, modem is not available, which is blocking the record
use case as record route is set to be MSIC Voice.
This workaround was introduced since android can start input stream
before changing the mode.

This patch switches the record in mode normal back to Multimedia route

Change-Id: I95565172c7b06baeca37f57c3d6b7edf55be8b18
Signed-off-by: Francois Gaffie <francoisx.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/37829
Reviewed-by: Soulard, MathieuX <mathieux.soulard@intel.com>
Reviewed-by: De Chivre, RenaudX <renaudx.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>