Audio-MW: [Audio]Sound cannot be heard after adjusting volume at settings on HDMI.
BZ: 31348

When "Broken_pipe" error occures while writing PCM data to driver, snd_pcm_recover() function is invoked.
This function returns error "-EBUSY" if the device is not ready, which causes this bug as this error is
not handled properly in AudioStreamOutALSA::write().
With this fix, error handleing mechanism is provided, which consumes the data for remaining duration before
invoking snd_pcm_recover() function and then returns "bytes" after PCM write recovery.

Change-Id: Ibafc622f194307b288f09bdfdf661bdc013813ce
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>
Reviewed-on: http://android.intel.com:8080/43896
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Hibare, PramodX <pramodx.hibare@intel.com>
Tested-by: Hibare, PramodX <pramodx.hibare@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>