[AUDIO][LEX] VoIP path moved to media route ie PCM2 MSIC
BZ: 31748

As Audience chip is not available on Lex, and as LPE needs to
be used to perform some processing instead, VoIP path must
be moved to Media Route, ie from IA I2S2 to MSIC PCM2. (This change
only affects when playing on Builtin accessories, not on BTSCO)

This patch changes the Voice and MSIC media routes applicability.

Change-Id: I0aef2604fc9c52b205edd0dfb32176dcb50120f0
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/44137
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>