[Audio_Comms] Music volume is too low during VOIP call
BZ: 32493

Nowadays music is too much attenuated during a VOIP call.
This is due to the IHF mode : without attenuation the music
level is too high when playing on speaker while calling on VOIP.
The solution is to put this attenuation in Audience but only on
speaker mode and not on earpiece or wired headset.

Change-Id: I8793f343cda97a61c1083b7bfb9ed15460c68148
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/44442
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>