[AUDIO] TTY does not work correctly
BZ: 31131

TTY does not work correctly. Advanced debug showed that audience
is not always in pass trough when TTY mode is on.
It is due to race condition as setParameter API of Audio HAL and
setStreamParameter can be called in any order to respectively
configure the tty mode and change the output device on jack insertion.

Audio HAL cannot rely on the right order. Need to reconsider the routing
in case of setting new tty mode to take it into accound.

Change-Id: Ifef4af5f95f0beb3c93b076bc77e6a65f8fb8591
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/45946
Reviewed-by: Schmuck, GuillaumeX <guillaumex.schmuck@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>