[Audio] Manage Audio diff between Medfield Tablet & Medfield Phone
BZ: 32826

The Medfield tablet does not support the ear piece device. This patch
updates the policy to handle a such case.

When the ear piece is not supported, the policy will use the speaker
instead.

The policy uses an Android property to check if the earpiece is
supported or not. The same mechanism is introduced for the back mic
which can be present or not.

By default (property not defined), the devices are supported. The
property must be explicitly defined to disable the device support.

Change-Id: I77cfbe11d33a4707e1dbd9b143812831b898f799
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/46712
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>