[REVERT] Audio-MW: [Audio]Sound cannot be heard after adjusting volume at settings on HDMI
BZ: 31926

After analysis and tests, the issue is a consequence of bug 31348 : if a
situation of broken pipe happens (now fixed by bug 33738), the phone
"plays" silence (sleep call), and tries to prepare the next buffer.
The result is silence then a fragment of valid audio played (again and again)
 till a new stream is played (system sound or tone for instance).

Solution consists in removing fix for 31348.

Change-Id: I13445fd56f32847bad1b82c740e7ce125a700611
Signed-off-by: guillaume schmuck <gschmucx@intel.com>
Reviewed-on: http://android.intel.com:8080/48344
Reviewed-by: Schmuck, GuillaumeX <guillaumex.schmuck@intel.com>
Reviewed-by: Di Folco, NeilX <neilx.di.folco@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>