Audio-MW: Sound cannot be heard after adjusting volume at settings on HDMI.
BZ: 31348

When EPIPE occurs in AudioStreamOutALSA::write(), snd_pcm_recover() function is invoked.
In case of HDMI, processing of this request is done at the interrupt boundary in driver code,
which would be responded by -EAGAIN till the interrupt boundary.
An error handling mechanism is provided with this fix, which sends repeated requests for
recovery after a delay of 10ms each time till the "totalSleepTime" is less than "Latency"
(i.e. "PERIOD_TIME * 2", for Playback in normal mode).

Change-Id: I4512c838fc127fd827539e170e09f3d7326c6b59
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>
Reviewed-on: http://android.intel.com:8080/49410
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Kandasamy, Muthukumar <muthukumar.kandasamy@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Reviewed-by: Schmuck, GuillaumeX <guillaumex.schmuck@intel.com>
Reviewed-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: Hibare, PramodX <pramodx.hibare@intel.com>
Tested-by: Hibare, PramodX <pramodx.hibare@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>