[AUDIO] Media server crash
BZ: 39345

Changes have been done within Audio HAL to have a cardinality
of 1 <> 1 between streams and ALSA handle. The stream must be deleted
first, then the ALSA device can be deleted (since the stream needs to
access its handle to unroute itself if not already done).
This could lead to a crash of a media server.

Change-Id: Ic836f98f25c372234c27ee328ab16eb1f0894916
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/50468
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>