[PORT FROM R3][AUDIO] Wrong Audience profile is set when connecting BT headset during call or communication
BZ: 38987

When connecting a BT headset during a voice call or communication, the AT+NREC command is sent
from BT headset to know if algorithms are processed in handset or BT headset. But it is sent
first with the ON parameter value, stating that processing occurs in handset (Audience). The
profile is applied and when the SCO link is really enabled for voice, AT+NREC=OFF command is
sent indicating that audio processing occurs in the BT headset. But as the audio routing has
already been done, the setting is not taken into account.
This patch reconsiders the routing so that the OFF setting is correctly applied by loading the
right Audience profile (in this case no_acoustic.bin).

Change-Id: I18bf23c55a256d09ed0ce5e73a04e37f6e36d376
Orig-Change-Id: Ib77db9fe68830e2b821575192f7c5f8cfc98ae15
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/53600
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>