[PORT FROM R3] [AUDIO] Crash in AT Manager
BZ: 40621

Even if the issue was raised with ENZO HW, the guilty code is
common and required to be patched.
When the modem is too long to answer (> client timeout), the
client semaphore is posted, and the client will return and send
quickly the following AT Command. The AT manager did not cancel
this request, leaving an orphan ATCmd pointer, that can lead to
a crash, even a seg fault.

The AT Manager has a protection mechanism: after 10 sec, if no
answer is received from the modem, it will reset the modem.
The timeout protection from the client is useless as the client
cannot recover from this error.

Note that for commodity reason, the recover mechanism of the
ATManager can be disabled using Android Property.

Change-Id: I2d9fe56865b928c7f67fd87d9142acfb229e0d06
Orig-Change-Id: Ib3e6b873c6ce6f4778ceaac10566683433e41e05
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/58516
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>