[PORT FROM R3][Audio] during CSV call, no uplink on Skype
BZ: 41012

There is an issue during CSV call : if the user makes a Skype call,
then there is no uplink on Skype.
Skype uses the BUILTIN_MIC source from the media path.
And during CSV call, the RouteManager cannot use this path.
The solution in AudioPolicy is to change the source into VOICE_CALL
so the uplink is mixed in the modem to the Skype application.
And we need also to change the inputSource in the HAL.

This is a port from R3-stable #53790 patch.

Change-Id: Ia253df0d6ac895ff1ac659847e1669a9f4f5706b
Origin-Change-Id: I581592f060b6f9141a187d0d886531d900492442
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/70278
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>