[PORT FROM R3] [AUDIO] Lost of 2 seconds of call when accepting it
BZ: 29679

The setMode in call was recently changed and is sent earlier.
The sequence is:
-setMode In call from the policy
-ModemCallStatus=true received from AT Manager
-setStreamParameter with the device to use received from
the policy.

As a result, when receiving the call status, the route is established
on the wrong device. When the setStreamParameter is received, a re-
routing is performed, increasing by around 800ms the establishment
of the audio call.

This patch reintroduce consistancy between the mode / device and
the routing, avoiding this extra re routing.

Change-Id: I10b4cade5b053bdef0e74444be367a5632c629e3
Orig-Change-Id: Ib2d01cc634757b095638e4b29fd22da0a97ceaa5
Old-Change-Id: I9839e95135e9433c1ddaa8abc994d3e7da957bf5
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/60263
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>