[PORT FROM R3] [AUDIO] No audio on the first call
BZ: 47641

This issue was seen on CTP, but it reveals a latent problem
within the common audio HAL code. The audio path is established
correctly, the volume is not set, remaining to 0.
In some case, the mode in call could be changed only after receiving
the modem audio status, changing the mode to the VPC and parameter
framework. The criteria was set, but the confirguration is not
applied...

Change-Id: I5edd370e8b5681411c95cc20c3489c5374c0d866
Orig-Change-Id: I5ae9459fef94f2b46300093643f97c575a6f87d3
Old-Change-Id: Ibb08e3e25cff60a16c921fba1e6b1ff6f12e7cba
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/60264
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>