[Audio] NB/WB support
BZ: 42137

Supporting narrow & wide bands Audience profile in VOIP mode implies
to know the VOIP audio signal sample rate.

This patch capture this information when the input stream for VOIP is
set and inform the VPC about it. VPC will be in charge to setup the
Audience audio algorithm & tuning for the VOIP band in use.

Change-Id: I59c9fadd1498a0cfea189f48fa7429ef55f5dff7
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/57738
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>