[PORT FROM R3] [AUDIO] Audio lost on BT after a CSV.
BZ: 48340

After a CSV call on BT, audio on BT is lost.
This is due to the accessibility of the BT route that has been
set to false to protect the I2S shared bus at the end of the call.
Accessibility must be restored.
This is a regression introduced by patch 54842.

This patch restores the reconsideration of the accessibility of the
route in case of Android Mode Change.
It also factorize for board without modem routes that have an
invariant accessibility.

Change-Id: Id70d1009b8a8ffbff46254be0f0eaa71e67dc3db
Orig-Change-Id: If7001979a42a81c3974aa63f078098f032b38bac
Old-Change-Id: I945e06a57eb6b2c8a0931c6bb4f452c51d2079a2
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/60265
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>