[AUDIO] Make the WiDi device take higher priority over BT device
BZ: 35312

The A/V is out of sync if Widi and Bluetooth are connected. This is because
audio is going to BT and video is going to Widi device. The video on Widi
has some latency and causes the audio on BT device to go out of sync.
This patch make the Widi device have a higher priority over BT device, so
that audio and video goes to Widi.

Change-Id: I70e68779d176935f408d37342337be2f6502f7e3
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/58398
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>