[AUDIO] SRC missing within Audio HAL
BZ: 49756

Some SRC are not supported by our resampler.
This is a fallback case, this first patch allows to use a pivot
resampler, ie a resampler to/from which all conversion are
supported.
This is a first dirty patch, better desgin to be done.

Final solution would be to enable all these SRC within
our resampler to avoid using 2 resamplers.

NOTE:
Resampler within HAL replaces ALSA plug conversion.
This conversion cannot be done into flinger in output, as flinger
does not change the sample specifications of the mixer thread
output.

Change-Id: I084a04cb9f1ae1941bbb29884b1d1ddf15965d5f
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/59913
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>