[AUDIO] Overwrite complete getInput method in derivated policy.
BZ: 49509

The base version of getInput method sets a channel mask in case of
input sources regarding voice call record (e.g. AUDIO_SOURCE_VOICE_CALL).
This channel mask info shall not be used in our platform as we are using
the input source to determine if record has to be done on one or two channels.

Change-Id: Icb970644fafb5b6d8b7b04e8484e69e6714eab23
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/59951
Reviewed-by: Bellouard, Jean-marie <jean-marie.bellouard@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>