[PORT FROM R3][AUDIO]Add new property modemIsPresent
BZ: 47366

Tablets support only HSP, and so cannot receive BT NREC commands. Therefore, acoustics should be always disabled when using BT on tablet, and NREC command support is removed from tablet.
This specific tablet behavior is managed through a new propertiy indicating if platform embeds a modem. This property is added to configuration files: if not specified, default is platform embeds a modem.
Note: This is a step forward on R4 as no tablet is running yet on this project, but this may happen in the future. The new property will be used anyway to remove use of CUSTOM_BOARD_WITHOUT_MODEM compilation flag.

Change-Id: I40a7ac8ec99586701665cefa564a5b42a4677579
Signed-off-by: abrabanx <aurelienx.braban@intel.com>
Reviewed-on: http://android.intel.com:8080/62994
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>