[Audio] Audio suspended while enabling/disabling HAC in-call
BZ: 53946

While in-call, when handset is not the current device, switching on
or off HAC should not produce anything. But audio is supended.

This is root caused by the unconditional rerouting which is done
each time HAC state changes. This patch force rerouting only if the
current device is earpiece (which means handset mode).

Change-Id: Ib66c0a41954c224797d13b4fac593055b44c4516
Signed-off-by: Mathieu Soulard <mathieu.soulard@intel.com>
Reviewed-on: http://android.intel.com:8080/63463
Reviewed-by: Braban, AurelienX <aurelienx.braban@intel.com>
Reviewed-by: Lee, JhinX <jhinx.lee@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>