[PORT FROM R3LEX][Audio HAL][Workaround]increase record buffersize
BZ: 54513

The patches of bug#33451 modified the bufferSize function
of ALSAStreamOps.cpp:
 - initially, following mode, the Audio HAL has been
configured at 20 ms if mode is normal or at 92 ms if mode
is in_call or in_communication;
 - in current implementation, if stream is out, it is
configured at 92 ms; if stream is input (record), it is
configured at 20 ms.
The purpose of this patch set is to increase to 80 ms the
bufferSize because at 92 ms, VoIP does not work and at
20 ms (as at 40 ms), Video recording does not work. Another
point: at 80 ms, it is probable that it improves our power
consummation.
This workaround will be replaced after investigation of
AudioMedia.

Change-Id: I3cce1a1bf7cb41f8eb235bd4dae10414f10d9c3d
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/63997
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>