[AUDIO] Remove AutoR lock in audio effects processing.
BZ: 56921

Remove call to AutoR on the lock object in echo reference methods. This allows
several clients to call this method at the same time. But in the call stack
of the same thread, an AutoW call (lock on write, only one client allowed to call
that method with the lock object) is done from AudioStreamInALSA.
Also, on the same lock object, setStreamParameters is called at the same time,
trying to acquire the lock with AutoW, blocking the mutex of the rwlock. And when
it comes to call AutoR, a deadlock happens as it is blocked, so removing it.

Change-Id: I55895dca5bcec3cb5f111f18254713dd39780df1
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/66514
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>