[WORKAROUND][Audio HAL] enable audio effects on output stream when applicable
BZ: 55434

The purpose of this patch is to control the usage of Android audio
effects. If either we don't have audience or we are not in a call,
then we can invoke Android audio effect, otherwise we rely on
Audience.
This patch does the same thing than made for BZ 31726, but for
output streams.

Change-Id: I80ab32c1ecb9c708494df49955a2678989608cea
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>
Reviewed-on: http://android.intel.com:8080/66789
Reviewed-by: Di Folco, NeilX <neilx.di.folco@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>