[AUDIO] Add a check when processing audio input frames.
BZ: 56473

Sometimes after processing, out_buf.frameCount is smaller than in_buf.frameCount,
as a result after several loops, processing_frames_in is already reduced to 0,
but the loop still continues. Adding a test in the while loop condition so that
the number of frames to be processed is always greater than 0 and we do not run
into an endless loop.

Change-Id: Icb8ac947645bdcf225ca00de3f51a8047a0117e6
Reviewed-on: http://android.intel.com:8080/66878
Reviewed-by: Becker, VincentX <vincentx.becker@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Jin, Zhi <zhi.jin@intel.com>
Reviewed-by: Xu, Even <even.xu@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>