[PORT FROM R3][AUDIO] VoIP cuts in UL
BZ: 57086

Some cut were still noticed in VoIP UL.
Pytimechart showed that the read of 80ms followed by the

WebRTC processing could delay the scheduling of the network
thread and we could miss the train.
This patch introduce the ability to change the buffering
modele on the input stream according to the input source
and set it to 20ms for the VoIP input source.

Change-Id: I7b2e097fa971e9383af0a1bd700e270e454a0002
orig-Change-Id: I0b368271a13e7db242973ad3a1c4ab8bd022461c
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/67950
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>