[WORKAROUND] [AUDIO] Set playback buffer time according to the sample rate
BZ: 56773

Period time is calculated now from the sample rate of the platform and
rounded up to a 16-frames boundary. This way, the size of the audio buffer
is always written at the right pace, and avoids possible underruns or overflows.
Use the resulting latency to calculate the buffer size for playback.

Change-Id: I9bcd1b23d63f06a079e32096efd30add56eee658
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Reviewed-on: http://android.intel.com:8080/67742
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>