[Audio] shutter sound and voice call in call+SCO
BZ: 59242

When taking a picture while in a call, the audio stream
is heard over the speakerphone for a second or two
when the shutter sound is played.

The solution is to set the device as BT SCO + Speaker.
The problem appears with Earpiece+Speaker as well.
ALSA will then recognize that it must be played as normal mode.
The Voice call will be unrouted and the shutter sound is played
on Speaker.
All the other cases (Widi, Wired Headset..) are set by the
fall through case "STARTEGY_MEDIA" in AudioPolicyManagerBase.

This patch is a port from patch 66797 from R3-stable.

Change-Id: I73fc213f78d9674f7f3757ce4df6f094913be69c
Signed-off-by: pafonsoX <philippex.afonso@intel.com>
Reviewed-on: http://android.intel.com:8080/68957
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>