[AudioHAL]memory init to avoid valgrind warning
BZ: 61412

To avoid a valgrind warning, it is necessary to initialize the
allocated memory in AudioStreamInAlsa. This initialization should
avoid also an unpredictable content of samples if the used buffer
is not entirely filled by HW; and 0 is more appropriate that
a random value from heap.
The issue has been found by valgrind.

Change-Id: I2c6c4aec2ce3fafca83a61d6cff6365227c51484
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/69511
Reviewed-by: Quintero, Jorge <jorge.quintero@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>