[AUDIO] Compilation error when enabling Audio PCM Dump from Android Audio HAL
BZ: 62456

Compilation error is due to the fact that the declaration and the initialisation
of audio_dump_filename is a scoping violation.
C++ standard ISO 14882:2003 says in §6.7.3:
It is possible to transfer into a block, but not in a way that bypasses declarations
with initialization.
A program that jumps from a point where a local variable with automatic storage
duration is not in scope to a point where it is in scope is ill-formed unless
the variable has POD type and is declared without an initializer

The solution here adopted is to limit the scope of the non-POD type within braces.
Another solution may be to move the previous ifcase with a goto statement below the ifdef section
Signed-off-by: pdiopx <piotrx.diop@intel.com>

Change-Id: Iabcb48796609a16a9dca8f05fa0f112df39c3236
Reviewed-on: http://android.intel.com:8080/70188
Reviewed-by: Diop, PiotrX <piotrx.diop@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Quervel, Thierry <thierry.quervel@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>