[AudioHAL]rework of remove_effect calls
BZ: 61296

Bug issue is probably (because unable to reproduce) caused by a
desynchronization of threads: a thread indirectly frees memory of
an "input stream" object then another thread continues to use
"intput stream" object. So, the purpose of this patch is
to add a lock before free of memory (which can increase the
robustness of AudioHAL to observed behaviour).

Change-Id: I2252ca50f14469c5059826d521e94b34f5dfd88a
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/71016
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>