[PORT FROM MAIN] GTI: start gti service in alsasound.
BZ: 66930

Having the GTI service to live in alsasound can be seen as a kind of workaournd.
Indeed, it real place would be in a real independent service. But for now, we do
not have enough time to implement this in AudioComms team and more over, the GTI
component is not under our responsability.

Change-Id: I8e8f1ac5c5e2be5b30c5c46cca082b73b1aab65f
Origin-Change-Id: Ife49ff6e7473fafc96698dc490ee717d9f351834
Signed-off-by: Selma Bensaid <selma.bensaid@intel.com>
Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>
Reviewed-on: http://android.intel.com:8080/78389
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>