[PORT FROM MAIN] [AUDIO] Grant MIC input creation while cam recording
BZ: 68227

This patch allows the creation of a concurrent input for
AUDIO_SOURCE_MIC while active camcorder recording session.
It allows to handle correctly an incoming GTalk voice call
while recording with the camera.
The camcorder recording session is stopped as soon as the
incoming call is ringing and not yet accepted nor decline.
This is the same behavior than ref phone (Galaxy nexus)

Change-Id: I3c42192078b3e19e18f7eb36ff03088a175ff713
Origin-Change-Id: Ic47ea13416122ad1e4909b75b019fedb5696d88a
Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>
Reviewed-on: http://android.intel.com:8080/78390
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>